+++ CHANGELOG.md
@@ -12,3 +12,6 @@ All Notable changes to `route-generator-plugin` will be documented in this file
 - Editor Config
 - PHP Unit distribution file
 - PHPUnit composer script
+- Route Generator
+- Exceptions
+- Tests